requirements.txt
streamlit>=1.32.0
pandas>=2.2.0
google-cloud-bigquery>=3.17.0
openpyxl>=3.1.2
db-dtypes>=1.2.0
pyarrow>=14.0.0
